refactor: FTH-95 sur screen, view model, and test
