+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/components/DotSeparator.kt
@@ -2,22 +2,24 @@ package net.thechance.mena.faith.presentation.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
 
 @Composable
-fun DotSeparator() {
+fun DotSeparator(
+    modifier: Modifier = Modifier,
+    color: Color = Theme.colorScheme.shadeTertiary
+) {
     Box(
-        modifier = Modifier
-            .padding(horizontal = Theme.spacing._8)
+        modifier = modifier
             .size(3.dp)
             .background(
-                color = Theme.colorScheme.shadeTertiary,
+                color = color,
                 shape = RoundedCornerShape(Theme.radius.full)
             )
     )
