+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/quran/bookmark/component/AyaBookmarkCard.kt
@@ -90,7 +90,7 @@ private fun SurahAndAyaInfo(
             style = Theme.typography.label.medium,
         )
 
-        DotSeparator()
+        DotSeparator(modifier = Modifier.padding(horizontal = Theme.spacing._8))
 
         Text(
             text = stringResource(Res.string.aya, ayaNumber),
