+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/quran/sur/SurInteractionListener.kt
@@ -1,9 +1,8 @@
 package net.thechance.mena.faith.presentation.feature.quran.sur
 
 interface SurInteractionListener {
-    fun onSurahClick(surahId: Int, surahName: String)
-    fun onBackClick()
-    fun onBookmarkClick()
-    fun onSearchClick()
-
+    fun onClickSurah(surahId: Int, surahName: String)
+    fun onClickBack()
+    fun onClickBookmark()
+    fun onClickSearch()
 }
